+++ src/lib.rs
@@ -35,7 +35,6 @@ mod vector;
 mod serialization;
 
 pub use angle::{Angle, AngleConsts, FloatAngleExt};
-use bytemuck::TransparentWrapper;
 pub use matrix::{Matrix2, Matrix3, Matrix4};
 pub use point::{Point2, Point3, Point4};
 pub use r#box::{Box2, Box3};
@@ -50,7 +49,9 @@ mod macros;
 use macros::*;
 
 #[doc(no_inline)]
-pub use traits::{Contains, Intersection, Union};
+pub use traits::{
+    peel, peel_mut, peel_ref, rewrap, wrap, Contains, Intersection, Transparent, Union,
+};
 
 /// Convenience glob import.
 ///
@@ -70,22 +71,6 @@ pub mod prelude {
 
     #[doc(no_inline)]
     pub use super::{point, point2, point3, point4, size, size2, size3, vec2, vec3, vec4, vector};
-}
-
-pub(crate) fn wrap<T, U: TransparentWrapper<T>>(t: T) -> U {
-    U::wrap(t)
-}
-pub(crate) fn peel<T, U: TransparentWrapper<T>>(u: U) -> T {
-    U::peel(u)
-}
-pub(crate) fn peel_ref<T, U: TransparentWrapper<T>>(u: &U) -> &T {
-    U::peel_ref(u)
-}
-pub(crate) fn peel_mut<T, U: TransparentWrapper<T>>(u: &mut U) -> &mut T {
-    U::peel_mut(u)
-}
-pub(crate) fn rewrap<A: TransparentWrapper<T>, B: TransparentWrapper<T>, T>(a: A) -> B {
-    B::wrap(A::peel(a))
 }
 
 /// Construct a [`Vector2`]. Usable in const contexts.